refactor(backend): change migration name
