+++ apps/backend/src/migrator/mod.rs
@@ -16,7 +16,7 @@ mod m20230814_add_lot_field_to_exercise;
 mod m20230814_add_muscles_field_to_exercise;
 mod m20230819_change_exercise_structure;
 mod m20230819_create_workout;
-mod m20230825_create_metadata_suggestion;
+mod m20230825_create_suggestion;
 
 pub use m20230410_create_metadata::{Metadata, MetadataImageLot, MetadataLot, MetadataSource};
 pub use m20230417_create_user::{UserLot, UserToMetadata};
@@ -48,7 +48,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20230814_add_muscles_field_to_exercise::Migration),
             Box::new(m20230819_change_exercise_structure::Migration),
             Box::new(m20230819_create_workout::Migration),
-            Box::new(m20230825_create_metadata_suggestion::Migration),
+            Box::new(m20230825_create_suggestion::Migration),
         ]
     }
 }
